changes to context menus
