+++ app/src/main/AndroidManifest.xml
@@ -27,8 +27,7 @@
         <activity
             android:name=".activities.ArchivedListActivity"
             android:label="@string/app_name"
-            android:parentActivityName=".activities.TaskListActivity"
-            android:windowSoftInputMode="adjustResize|stateAlwaysVisible" />
+            android:parentActivityName=".activities.TaskListActivity" />
     </application>
 
 </manifest>
