+++ app/src/main/java/com/rahobbs/todo/fragments/ArchivedListFragment.java
@@ -8,11 +8,16 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.rahobbs.todo.R;
-import com.rahobbs.todo.activities.TodoPagerActivity;
+import com.rahobbs.todo.activities.ArchivedListActivity;
+import com.rahobbs.todo.helpers.Feedback;
+import com.rahobbs.todo.helpers.SharableList;
 import com.rahobbs.todo.helpers.TodoItem;
 import com.rahobbs.todo.helpers.TodoLab;
 import com.rahobbs.todo.interfaces.SimpleItemTouchHelperCallback;
@@ -32,6 +37,7 @@ public class ArchivedListFragment extends TodoListFragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        context_menu = R.menu.archive_context_menu;
         setHasOptionsMenu(true);
     }
 
@@ -75,6 +81,32 @@ public class ArchivedListFragment extends TodoListFragment {
             mTodoRecyclerView.setAdapter(new TodoListAdapter(todoItems));
         } else {
             recyclerAdapter.setTodoItems(todoItems);
+        }
+    }
+
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        super.onCreateOptionsMenu(menu, inflater);
+
+        inflater.inflate(R.menu.fragment_archive_list, menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.share_list:
+                List<TodoItem> list = TodoLab.get(getActivity()).getUnarchivedItems();
+                SharableList sl = new SharableList();
+                startActivity(Intent.createChooser(sl.makeSharable(list), "Send list..."));
+                return true;
+            case R.id.menu_item_send_feedback:
+                Intent i = new Intent(Intent.ACTION_SEND);
+                Feedback fb = new Feedback();
+                fb.sendFeedback(i);
+                startActivity(Intent.createChooser(i, "Send mail..."));
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
         }
     }
