+++ app/src/main/res/menu/archive_context_menu.xml
@@ -14,12 +14,6 @@
         app:showAsAction="ifRoom"/>
 
     <item
-        android:id="@+id/share_list"
-        android:title="Share List"
-        app:showAsAction="ifRoom"
-        android:icon="@drawable/ic_share_list"/>
-
-    <item
         android:id="@+id/close_menu"
         android:title="Close"
         android:icon="@drawable/ic_close_menu"
