+++ app/src/main/res/menu/context_menu.xml
@@ -12,11 +12,6 @@
         android:title="Delete Items"
         android:icon="@drawable/ic_menu_delete"
         app:showAsAction="ifRoom"/>
-    <item
-        android:id="@+id/share_list"
-        android:title="Share List"
-        app:showAsAction="ifRoom"
-        android:icon="@drawable/ic_share_list"/>
 
     <item
         android:id="@+id/close_menu"
