+++ app/src/main/res/menu/fragment_archive_list.xml
@@ -1,19 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
-      xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
         android:id="@+id/share_list"
         android:title="Share List"
         app:showAsAction="always"
         android:icon="@drawable/ic_share_list"/>
     <item
-        android:id="@+id/menu_item_view_archive"
-        android:title="View archive"
-        app:showAsAction="never"
-        />
-    <item
         android:id="@+id/menu_item_send_feedback"
         android:title="@string/send_feedback"
         app:showAsAction="never"/>
-
 </menu>
